refactor(chat): extract ask-user-question renderer from ToolCallRenderer
ToolCallRenderer is pure free functions; the ask-user-question rendering cluster
(formatAnswer/resolveAskUserAnswers/renderAskUserQuestionResult/Fallback/Option)
moves to askUserQuestionRenderer.ts, and the shared contentFallback primitive to
its own module so renderer siblings can reuse it without a cycle.

Behavior-preserving (ToolCallRenderer suite — 175 tests — passes untouched
through the public renderToolContent/updateToolCallResult surface).
ToolCallRenderer 854 -> 766 LOC.
