+++ src/features/chat/rendering/ToolCallRenderer.ts
@@ -2,7 +2,6 @@ import { type App,setIcon } from 'obsidian';
 
 import type { TodoItem } from '../../../core/tools/todo';
 import { getToolIcon, MCP_ICON_MARKER } from '../../../core/tools/toolIcons';
-import { extractResolvedAnswersFromResultText } from '../../../core/tools/toolInput';
 import {
   isAgentLifecycleTool,
   TOOL_APPLY_PATCH,
@@ -24,7 +23,7 @@ import {
   TOOL_WRITE_STDIN,
 } from '../../../core/tools/toolNames';
 import { extractToolResultContent } from '../../../core/tools/toolResultContent';
-import type { AskUserQuestionItem, AskUserQuestionOption, ToolCallInfo } from '../../../core/types';
+import type { ToolCallInfo } from '../../../core/types';
 import type { DiffStats } from '../../../core/types/diff';
 import { appendMcpIcon } from '../../../shared/icons';
 import { parseApplyPatchDiffs, parseFileUpdateChangeDiffs } from '../../../utils/diff';
@@ -34,7 +33,12 @@ import {
   renderApplyPatchChangeList,
   renderApplyPatchResultFallback,
 } from './applyPatchExpandedHelpers';
+import {
+  renderAskUserQuestionFallback,
+  renderAskUserQuestionResult,
+} from './askUserQuestionRenderer';
 import { setupCollapsible } from './collapsible';
+import { contentFallback } from './contentFallback';
 import { renderDiffContent, renderDiffStats } from './DiffRenderer';
 import { renderTodoItems } from './todoUtils';
 import {
@@ -678,117 +682,6 @@ function createToolElementStructure(
   const content = toolEl.createDiv({ cls: 'specorator-tool-content' });
 
   return { toolEl, header, iconEl, nameEl, summaryEl, statusEl, content, currentTaskEl };
-}
-
-function formatAnswer(raw: unknown): string {
-  if (Array.isArray(raw)) return raw.join(', ');
-  if (typeof raw === 'string') return raw;
-  return '';
-}
-
-function resolveAskUserAnswers(toolCall: ToolCallInfo): Record<string, unknown> | undefined {
-  if (toolCall.resolvedAnswers) return toolCall.resolvedAnswers;
-
-  const parsed = extractResolvedAnswersFromResultText(toolCall.result);
-  if (parsed) {
-    toolCall.resolvedAnswers = parsed;
-    return parsed;
-  }
-
-  return undefined;
-}
-
-function renderAskUserQuestionResult(container: HTMLElement, toolCall: ToolCallInfo): boolean {
-  container.empty();
-  const questions = toolCall.input.questions as AskUserQuestionItem[] | undefined;
-  const answers = resolveAskUserAnswers(toolCall);
-  if (!questions || !Array.isArray(questions) || !answers) return false;
-
-  const reviewEl = container.createDiv({ cls: 'specorator-ask-review' });
-  for (let i = 0; i < questions.length; i++) {
-    const q = questions[i];
-    const answer = formatAnswer(
-      (q.id ? answers[q.id] : undefined) ?? answers[q.question]
-    );
-    const pairEl = reviewEl.createDiv({ cls: 'specorator-ask-review-pair' });
-    pairEl.createDiv({ text: `${i + 1}.`, cls: 'specorator-ask-review-num' });
-    const bodyEl = pairEl.createDiv({ cls: 'specorator-ask-review-body' });
-    bodyEl.createDiv({ text: q.question, cls: 'specorator-ask-review-q-text' });
-    bodyEl.createDiv({
-      text: answer || 'Not answered',
-      cls: answer ? 'specorator-ask-review-a-text' : 'specorator-ask-review-empty',
-    });
-  }
-
-  return true;
-}
-
-function renderAskUserQuestionFallback(container: HTMLElement, toolCall: ToolCallInfo, initialText?: string): void {
-  container.empty();
-
-  const questions = Array.isArray(toolCall.input.questions)
-    ? toolCall.input.questions as AskUserQuestionItem[]
-    : [];
-
-  if (questions.length === 0) {
-    contentFallback(container, initialText || toolCall.result || 'Waiting for answer...');
-    return;
-  }
-
-  if (initialText || toolCall.result) {
-    container.createDiv({
-      cls: 'specorator-ask-review-prompt',
-      text: initialText || toolCall.result || 'Waiting for answer...',
-    });
-  }
-
-  for (let questionIndex = 0; questionIndex < questions.length; questionIndex++) {
-    const question = questions[questionIndex];
-    const reviewEl = container.createDiv({ cls: 'specorator-ask-review' });
-    const pairEl = reviewEl.createDiv({ cls: 'specorator-ask-review-pair' });
-    pairEl.createDiv({ text: `${questionIndex + 1}.`, cls: 'specorator-ask-review-num' });
-    const bodyEl = pairEl.createDiv({ cls: 'specorator-ask-review-body' });
-    bodyEl.createDiv({ text: question.question, cls: 'specorator-ask-review-q-text' });
-
-    if (!Array.isArray(question.options) || question.options.length === 0) {
-      bodyEl.createDiv({ cls: 'specorator-ask-review-empty', text: 'No options recorded' });
-      continue;
-    }
-
-    const listEl = bodyEl.createDiv({ cls: 'specorator-ask-list' });
-    question.options.forEach((option, optionIndex) => {
-      renderAskUserQuestionOption(listEl, option, optionIndex, question.multiSelect === true);
-    });
-  }
-}
-
-function renderAskUserQuestionOption(
-  parentEl: HTMLElement,
-  option: AskUserQuestionOption,
-  optionIndex: number,
-  isMultiSelect: boolean,
-): void {
-  const itemEl = parentEl.createDiv({ cls: 'specorator-ask-item is-disabled' });
-
-  if (isMultiSelect) {
-    itemEl.createDiv({ cls: 'specorator-ask-check', text: '[ ] ' });
-  } else {
-    itemEl.createDiv({ cls: 'specorator-ask-item-num', text: `${optionIndex + 1}. ` });
-  }
-
-  const contentEl = itemEl.createDiv({ cls: 'specorator-ask-item-content' });
-  const labelRowEl = contentEl.createDiv({ cls: 'specorator-ask-label-row' });
-  labelRowEl.createDiv({ cls: 'specorator-ask-item-label', text: option.label });
-
-  if (option.description) {
-    contentEl.createDiv({ cls: 'specorator-ask-item-desc', text: option.description });
-  }
-}
-
-function contentFallback(container: HTMLElement, text: string): void {
-  const resultRow = container.createDiv({ cls: 'specorator-tool-result-row' });
-  const resultText = resultRow.createSpan({ cls: 'specorator-tool-result-text' });
-  resultText.setText(text);
 }
 
 function renderBashContent(
